Remove test_util.h, move functions to test.h
